net/mlx5: E-Switch, Restore chain id on miss
We mark packets that miss on some chain with the current chain id
via regiser C, this register survives loopback and gets to
the CQE, we then use it to tell tc to continue in the
sw chain where we left off via the tc skb extension.

Chain ids are mapped to the lower part of reg C, and after loopback are
copied to to CQE via a restore rule's flow_tag. For each chain id there
would be one such restore rule.

Issue: 1806821
Change-Id: Ia15ebc708b9095fc768f367a2072d15928f1974d
Signed-off-by: Paul Blakey <paulb@mellanox.com>