net/mlx5e: tc_ct: Add tc connection tracking support
Add new module en/tc_ct that will be in charge of offloading
tc connection tracking rules.

The connection tracking rules aren't directly offloaded, but instead
we save them in a table and report to tc that they are offloaded.
So packets will not match this rule in hardware and go through software.

We then listen for packets going though the software rules in flower, and
check their connection tracking status in their zone, if they are in an established state,
We offload merged rules with a match on this connection tuple and all original non-offloaded
rules (but without the action ct) corrosponding to that conntrack zone.

For this offloaded merged rules, we also add a match on tcp_flags=-fin-syn-rst+ack,
so this packets that can change the connection state, go through software.
We also add a header rewrite action to mark them with a tracked and established
state and their zone via registers.

The ct_state and ct_zone matches will be translated to this registers.

We also do aging for this offloaded merged rules and remove them
after a configured time.

Issue: 1806821
Change-Id: I914b14a342645dbefb61e75f65a8319c1bb932eb
Signed-off-by: Paul Blakey <paulb@mellanox.com>