net/mlx5e: en_tc: Support inner header rewrite with goto action
Currently, Hardware supports header rewrite of outter headers only.
So to use header rewrite on inner headers, one must first
decapsulate the packet. Since we translate tunnel_key unset to hardware
decap action, this won't work with goto action as matches on the next
chain will be wrongly translated to match on inner headers which no
longer exists.

To support header rewrite for inner with goto action, eswitch flows
with goto actions on chain 0 will implictily decapsulate the packet, and
set a metadata mapping for the decapuslated header to match against in
later chains matces.

As this is a mapping, later chains tc flows must match on the same
tunnel matching to be supported, as they are translated to a match
on the saved metadata.

We also be using this mapping to restore the tunnel info metadata
on miss.

Issue: 1877135
Change-Id: Iba6ee240d46c5ad4d844992eeb13d4bee016f2c7
Signed-off-by: Paul Blakey <paulb@mellanox.com>