net/mlx5: E-Switch, Restore tunnel metadata on miss
Since we decapsulate the pakcets on first chain, if we miss on later
chains, the packet won't be taken by the vxlan device automatically.
So we reinject it on the vxlan device from the current tunnel id
(which we set on the implicit decap on chain 0), and restore the tunnel
metadata so further tc matches would work.

Issue: 1877135
Change-Id: I473a92a9a8350fe8c98cd4ad05b7a214cb24e4eb
Signed-off-by: Paul Blakey <paulb@mellanox.com>